feat(ux): show next-review interval on grade buttons and in card list
- Study screen (web + Android): each grade button now shows the
  projected interval beneath the label (e.g. "Hard / 3d") so the user
  knows exactly where the card will land before tapping.
- Deck detail (web + Android): each card row shows "↻ <when>" — how
  long until the next review, relative to now (sub-day: min/h; longer:
  d / mo / y).

fix(backend): align ease_factor floor with float32 lattice

Android stores ease_factor as a 32-bit Float, so the real SM-2 floor
is float32(1.3) = 1.2999999523…, not the double 1.3. The old backend
validation rejected a legitimately-floored card from the web client
with 400 "ease_factor out of range". Align all three surfaces:
  • web sm2.ts   EASE_MIN = Math.fround(1.3)
  • backend sm2.go  minEaseFactor = float64(float32(1.3))
  • Android already uses 1.3f — no change needed.

test: cross-platform SM-2 parity suite against shared golden fixtures

Add shared/testdata/{sm2_vectors,interval_vectors}.json generated from
the web reference implementation (npm run test:gen). Both clients assert
against the same files — passing on both sides proves the SM-2 algorithm
and interval formatting are bit-identical across platforms.
  • web/test/parity.test.ts   (node --test, no new deps)
  • CrossPlatformParityTest.kt (Gson, existing dep)
The parity tests are what caught the float32 ease divergence above.

Also add a safety guard to integration_helpers_test.go: refuses to
TRUNCATE unless current_database() == "testdb", protecting prod DB if
TEST_DATABASE_URL is accidentally pointed at it.

+++ backend/internal/handler/sm2.go
@@ -9,9 +9,15 @@ import (
 // SM-2 field bounds, shared by the strict PUT /cards/{id} validation and the
 // lenient /sync sanitisation so both paths agree on what a storable card is.
 const (
-	minInterval    = 1
-	maxInterval    = math.MaxInt32 // backend stores interval as a 32-bit INTEGER
-	minEaseFactor  = 1.3
+	minInterval = 1
+	maxInterval = math.MaxInt32 // backend stores interval as a 32-bit INTEGER
+	// The clients compute ease_factor on a float32 lattice (Android stores it as
+	// a 32-bit Float; web matches via Math.fround). Their floor is therefore
+	// float32(1.3) = 1.2999999523…, not the double 1.3. Using 1.3 here would
+	// reject a legitimately-floored card on strict PUT (400) and clamp it upward
+	// on every /sync — churning updated_at and re-diverging the clients. Keep the
+	// bound on the same lattice so a floored card round-trips untouched.
+	minEaseFactor  = float64(float32(1.3)) // 1.2999999523162842
 	maxEaseFactor  = 5.0
 	minRepetitions = 0
 	maxRepetitions = math.MaxInt32 // repetitions is a 32-bit INTEGER column too
